refactor: move edit modal to its own file
